ext/wavpack/gstwavpackparse.*: Always report the duration if we know it in push mode and don't return 0 just to make ...
Original commit message from CVS:
* ext/wavpack/gstwavpackparse.c: (gst_wavpack_parse_src_query),
(gst_wavpack_parse_create_src_pad):
* ext/wavpack/gstwavpackparse.h:
Always report the duration if we know it in push mode and don't
return 0 just to make totem believe we can't seek in push mode.
Newer totem version use the SEEKING query which properly reports
if we can seek or not.

+++ ext/wavpack/gstwavpackparse.h
@@ -57,7 +57,7 @@ struct _GstWavpackParse
 
   guint samplerate;
   guint channels;
-  guint total_samples;
+  gint64 total_samples;
 
   gboolean need_newsegment;
   gboolean discont;
